Move line ending input streams to normalization-java
