+++ subprojects/normalization-java/src/main/java/org/gradle/api/internal/changedetection/state/LineEndingNormalizingInputStream.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 the original author or authors.
+ * Copyright 2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.hash;
+package org.gradle.api.internal.changedetection.state;
 
 import com.google.common.primitives.Bytes;
